fix: compile for Postgres + harden deploy gate
- pagination limit/offset u32 -> i64 so sqlx binds for Postgres (models.rs, shared.rs, admin.rs)

- enable chrono 'serde' feature for DateTime<Utc> serialization

- merge flaky env-var orchestrator_timeout tests into one (no parallel race)

- Dockerfile: run as non-root appuser (SOC2 CC6.8)

- deploy gate: scope tests to lib/bins (integration suite needs Postgres), add non-blocking cargo audit

+++ src/lib/models.rs
@@ -45,8 +45,8 @@ pub struct UpdateTaskRequest {
 
 #[derive(Debug, Deserialize, Default)]
 pub struct ListTasksQuery {
-    pub limit: Option<u32>,
-    pub offset: Option<u32>,
+    pub limit: Option<i64>,
+    pub offset: Option<i64>,
     pub completed: Option<bool>,
     pub status: Option<String>,
     pub q: Option<String>,
@@ -129,8 +129,8 @@ pub struct AdminRequestLog {
 
 #[derive(Debug, Deserialize, Default)]
 pub struct AdminListQuery {
-    pub limit: Option<u32>,
-    pub offset: Option<u32>,
+    pub limit: Option<i64>,
+    pub offset: Option<i64>,
 }
 
 #[derive(Debug, Serialize, FromRow)]
